+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## Unreleased
+- Added new User Agent profiles (Firefox, Edge, Samsung Internet).
+- Added a Custom User Agent profile with user-defined string.